# demo_ecommerce_api/openapi.yaml
openapi: 3.1.0
info:
  title: E-Commerce Store API
  version: 1.0.0
  description: API for products, orders, and customer management.
servers:
- url: http://localhost:8000
  description: Local Development Server
paths:
  /products:
    get:
      operationId: get_products
      summary: List products
      description: List products
      tags:
      - Products
      parameters:
      - name: limit
        in: query
        required: false
        description: Maximum number of records to return
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      - name: offset
        in: query
        required: false
        description: Number of records to skip for pagination
        schema:
          type: integer
          default: 0
          minimum: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Bad Request / Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: post_products
      summary: Create product
      description: Create product
      tags:
      - Products
      parameters: []
      responses:
        '201':
          description: Resource created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Bad Request / Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized / Missing Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        description: Payload data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
  /products/{product_id}:
    get:
      operationId: get_products_product_id
      summary: Get product
      description: Get product
      tags:
      - Products
      parameters:
      - name: product_id
        in: path
        required: true
        description: Unique identifier for product_id
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Bad Request / Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders:
    post:
      operationId: post_orders
      summary: Place order
      description: Place order
      tags:
      - Orders
      parameters: []
      responses:
        '201':
          description: Resource created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad Request / Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized / Missing Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        description: Payload data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
    get:
      operationId: get_orders
      summary: List orders
      description: List orders
      tags:
      - Orders
      parameters:
      - name: limit
        in: query
        required: false
        description: Maximum number of records to return
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      - name: offset
        in: query
        required: false
        description: Number of records to skip for pagination
        schema:
          type: integer
          default: 0
          minimum: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          description: Bad Request / Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized / Missing Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /orders/{order_id}:
    get:
      operationId: get_orders_order_id
      summary: Get order
      description: Get order
      tags:
      - Orders
      parameters:
      - name: order_id
        in: path
        required: true
        description: Unique identifier for order_id
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad Request / Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized / Missing Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
components:
  schemas:
    ErrorResponse:
      type: object
      required:
      - detail
      - status
      properties:
        detail:
          type: string
          description: Human-readable error description
          example: sample_detail
        status:
          type: integer
          description: HTTP status code
          example: 101
        code:
          type: string
          description: Error classification code
          example: sample_code
      example:
        detail: sample_detail
        status: 101
        code: sample_code
    Product:
      type: object
      description: Store merchandise
      required:
      - id
      - title
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        title: &id001
          type: string
          example: Sample title
        price: &id002
          type: string
          example: Sample price
        stock: &id003
          type: integer
          example: 1
      example:
        id: 123e4567-e89b-12d3-a456-426614174000
        title: Sample title
        price: Sample price
        stock: 1
    ProductCreate:
      type: object
      description: Payload for creating a new Product
      required:
      - title
      properties: &id004
        title: *id001
        price: *id002
        stock: *id003
      example:
        title: Sample title
        price: Sample price
        stock: 1
    ProductUpdate:
      type: object
      description: Payload for updating an existing Product
      properties: *id004
      example:
        title: Sample title
        price: Sample price
        stock: 1
    ProductListResponse:
      type: object
      description: Paginated list of Product records
      required:
      - items
      - total
      - limit
      - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
          example: []
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
      example:
        items: []
        total: 100
        limit: 20
        offset: 0
    Order:
      type: object
      description: Customer purchase order
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        customer_id: &id005
          type: string
          example: Sample customer_id
        total_amount: &id006
          type: string
          example: Sample total_amount
        status: &id007
          type: string
          example: Sample status
      example:
        id: 123e4567-e89b-12d3-a456-426614174000
        customer_id: Sample customer_id
        total_amount: Sample total_amount
        status: Sample status
    OrderCreate:
      type: object
      description: Payload for creating a new Order
      required: []
      properties: &id008
        customer_id: *id005
        total_amount: *id006
        status: *id007
      example:
        customer_id: Sample customer_id
        total_amount: Sample total_amount
        status: Sample status
    OrderUpdate:
      type: object
      description: Payload for updating an existing Order
      properties: *id008
      example:
        customer_id: Sample customer_id
        total_amount: Sample total_amount
        status: Sample status
    OrderListResponse:
      type: object
      description: Paginated list of Order records
      required:
      - items
      - total
      - limit
      - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
          example: []
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
      example:
        items: []
        total: 100
        limit: 20
        offset: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token